Add cloud build support
In order for all these fun moving parts to work, I needed to shift us to the 2.0 release candidate
for the App Engine maven plugin.  This translates to a few changes to how we run things for the project

mvn appengine:start  # to start up your dev server
mvn appengine:stop   # to stop it.

Once this is committed, we can enable a build hook against the master branch that automatically builds
and deploys our service.